chore(blueprint): Organize blueprint tests
Splits blueprint tests in to "public" and "private" tests, shortening the number of lines per test file substantially.

Also made several tests more compliant with style guideilnes. Coverage increased slightly.
